// src/app/models/event.ts
export interface Event {
    id: number;
    title: string;
    partner: string;
    start: Date;
    color: string;
    status: string;
    insurance: string;
    resourceId: string;
    addtitle1: string;
    addtitle2: string;
    addtitle3: string;
    comment: string;
    genericChar1: string;
    genericChar2: string;
    genericChar3: string;
    genericTime1: Date;
    genericNumber1: number;
    genericNumber2: number;
    genericNumber3: number;
    persona: number; 
    kollege: number;
}